// adopt-me-typescript/src/typings/ApiResponseTypes.ts
export type Animal = 'dog' | 'cat' | 'bird' | 'reptile' | 'rabbit' | '';
export interface Pet {
  id: number;
  name: string;
  animal: Animal;
  description: string;
  breed: string;
  images: string[];
  city: string;
  state: string;
}

export interface PetApiResponse {
  numberOfResults: number;
  startIndex: number;
  endIndex: number;
  hasNext: boolean;
  pets: Pet[];
}

export interface AnimalListAPIResponse {
  animal: Animal;
  breeds: string[];
}